client: fix UI not updating according to grants
